Refactor: 식단 관련 엔티티 수정
