--- src/main/java/calocheck/boundedContext/dailyFoodInfo/repository/DailyFoodInfoRepository.java
@@ -1,10 +1,0 @@
-package calocheck.boundedContext.dailyFoodInfo.repository;
-
-import calocheck.boundedContext.dailyFoodInfo.entity.DailyFoodInfo;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DailyFoodInfoRepository extends JpaRepository<DailyFoodInfo, Long> {
-
-
-
-}
