+++ src/main/java/calocheck/boundedContext/dailyMenu/service/DailyMenuService.java
@@ -1,50 +1,88 @@
 package calocheck.boundedContext.dailyMenu.service;
 
-import calocheck.boundedContext.dailyFoodInfo.entity.DailyFoodInfo;
+import calocheck.boundedContext.cartFoodInfo.entity.CartFoodInfo;
+import calocheck.boundedContext.cartFoodInfo.service.CartFoodInfoService;
 import calocheck.boundedContext.dailyMenu.entity.DailyMenu;
 import calocheck.boundedContext.dailyMenu.repository.DailyMenuRepository;
 import calocheck.boundedContext.member.entity.Member;
+import calocheck.boundedContext.nutrient.entity.Nutrient;
+import calocheck.boundedContext.nutrient.service.NutrientService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class DailyMenuService {
-    private final DailyMenuRepository dailyMenuRepository;
 
-//    public DailyMenu create(Member member, List<DailyFoodInfo> dailyFoodInfos) {
-//        DailyMenu dailyMenu = DailyMenu.builder()
-//                .member(member)
-//                .dailyFoodInfoList(dailyFoodInfos)
+    private final DailyMenuRepository dailyMenuRepository;
+    private final CartFoodInfoService cartFoodInfoService;
+    private final NutrientService nutrientService;
+
+    public void create(Member member, String mealTime, int menuScore, String menuMemo) {
+
+        List<CartFoodInfo> cartByMember = cartFoodInfoService.findAllByMember(member);
+
+
+
+    }
+
+//    //식단 수정? => 가능한가? 어디에서?
+//    public DailyFoodInfo udpate(DailyFoodInfo dailyFoodInfo) {
+//
+//        DailyFoodInfo updated = dailyFoodInfo.toBuilder()
+//                .mealTime(meal)
 //                .build();
 //
-//        return dailyMenuRepository.save(dailyMenu);
+//        return dailyFoodInfoRepository.save(updated);
 //    }
-//
-//    public DailyMenu udpate(DailyMenu dailyMenu, Member member, List<DailyFoodInfo> dailyFoodInfos) {
-//        DailyMenu updated = dailyMenu.toBuilder()
-//                .member(member)
-//                .dailyFoodInfoList(dailyFoodInfos)
-//                .build();
-//
-//        return dailyMenuRepository.save(updated);
-//    }
-//
-//    public void delete(DailyMenu dailyMenu) {
-//        dailyMenuRepository.delete(dailyMenu);
-//    }
-//
-//    public DailyMenu findById(Long id) {
-//        Optional<DailyMenu> dailyMenu = dailyMenuRepository.findById(id);
-//
-//        return dailyMenu.orElse(null);
-//    }
-//
-//    public DailyMenu getByMember(Member member){
-//
-//        return dailyMenuRepository.findByMember(member).orElse(null);
-//    }
+
+    public void delete(DailyMenu dailyMenu) {
+        dailyMenuRepository.delete(dailyMenu);
+    }
+
+    public DailyMenu findById(Long id) {
+        Optional<DailyMenu> dailyFoodInfo = dailyMenuRepository.findById(id);
+
+        return dailyFoodInfo.orElse(null);
+    }
+
+    //권장량 - total => (권장량 - 오늘 먹은 양) - 지금 먹을 양
+    public List<Nutrient> calcNutrient(Member member, List<Nutrient> cartNutrientTotal) {
+
+        //1. 권장량
+        //2. 오늘 먹은 양
+        //3. 지금 먹을 양
+        //(권장량 - 오늘먹은 양) - 지금 먹을 양
+
+        //FIXME Sample
+        String[] nutrientName = {"단백질", "탄수화물", "지방"};
+
+        List<Nutrient> calcList = Arrays.stream(nutrientName)
+                .map(name -> cartNutrientTotal.stream()
+                        .filter(nutrient -> nutrient.getName().equals(name))
+                        .findFirst()
+                        .map(nutrient -> Nutrient.builder()
+                                .name(name)
+                                .value(formattingDoubleValue(nutrient.getValue() - 20 - 10)) // FIXME (권장량 - 오늘먹은 양) - 지금 먹을 양
+                                .unit(nutrient.getUnit())
+                                .build())
+                        .orElse(null))
+                .filter(Objects::nonNull)
+                .collect(Collectors.toList());
+
+        return calcList;
+    }
+
+    public double formattingDoubleValue(double value) {
+
+        BigDecimal decimal = BigDecimal.valueOf(value);
+        decimal = decimal.setScale(2, RoundingMode.HALF_UP);
+
+        return decimal.doubleValue();
+    }
 }
